Demonstrate problem with reference creation that can fail
Say the `Command` to create a new `Reference` can fail undeterministically,
e.g., the corresponding `Response` contains `Either
String (Reference (Opaque (IORef Int) r)`. With 'undeterministically', I mean
that it cannot be predicted based on the current symbolic model.

This is problematic for quickcheck-state-machine and results in an
exception (shown below). The mocked 'Response' to this 'Command' does not know
whether to fail or not, and will in some cases return a different number of
references than the actual response (1 vs. 0). In `insertConcretes`, the
symbolic reference variables are paired up concrete reference variables, and a
mismatch here results in the exception below.

To demonstrate this problem, I modified the `MemoryReference` example so that
`Create` randomly fails. Use `quickCheck (prop_sequential None)` to produce
the failure below:

    *** Failed! Falsifiable (after 5 tests and 1 shrink):
    Commands
      { unCommands =
          [ Command Create (fromList [ Var 0 ])
          , Command (Read (Reference (Symbolic (Var 0)))) (fromList [])
          ]
      }

    Model []

       == Create ==> Created (Left "newIORef failed") [ 0 ]
    *** Exception running callback:
    insertConcrets: impossible.
    CallStack (from HasCallStack):
      error, called at src/Test/StateMachine/Types/Environment.hs:73:3 in main:Test.StateMachine.Types.Environment